// src/types/menuKeyType.ts
export type MenuKey =
  | "dashboard"
  | "gestao_contas"
  | "transferencias"
  | "pagamentos"
  | "depositos_gpo"
  | "levantamentos"
  | "movimentos"
  | "relatorios"
  | "logs"
  | "saldo_por_usuario"
  | "permissoes"
  | "usuarios_backoffice"
  | "servicos"
  | "bancos"
  | "campanhas"
  | "promocoes"
  | "publicidades"
  | "faqs"
  | "meu_perfil"
  | "gestao_de_niveis"